update list command and tests
