+++ cli/package.json
@@ -1,7 +1,7 @@
 {
 	"name": "@gabimoncha/cursor-rules",
 	"description": "A CLI for bootstrapping Cursor rules to a project",
-	"version": "0.1.9",
+	"version": "0.2.0",
 	"type": "module",
 	"main": "./lib/index.js",
 	"types": "./lib/index.d.ts",
@@ -60,7 +60,7 @@
 		"package-manager-detector": "^1.3.0",
 		"picocolors": "^1.0.1",
 		"regex": "^6.0.1",
-		"repomix": "^0.3.9",
+		"repomix": "^1.0.0",
 		"semver": "^7.7.2",
 		"zod": "^3.25.67"
 	},
