+++ cli/src/cli/actions/repomixAction.ts
@@ -1,3 +1,4 @@
+import { spawn, fork } from 'node:child_process';
 import { writeFileSync } from 'node:fs';
 import path from 'node:path';
 import pc from 'picocolors';
