+++ example/package.json
@@ -3,11 +3,6 @@
 	"version": "0.0.1",
 	"private": true,
 	"scripts": {
-		"rules:dev": "bun run --bun cursor-rules",
-		"rules:lib": "cursor-rules",
 		"clean": "rm -rf repomix-output.xml repomix.config.json && find .cursor/rules -mindepth 1 ! -name 'bad-rule.mdc' -exec rm -rf -- {} +"
-	},
-	"devDependencies": {
-		"@gabimoncha/cursor-rules": "workspace:*"
 	}
 }
